+++ StudentSoup/src/main/frontend/src/components/boards/BoardBestReviewHeart.tsx
