+++ StudentSoup/src/main/frontend/src/components/boards/BoardWrite.tsx
