+++ StudentSoup/src/main/frontend/src/components/boards/boardBestReviewHeart.scss
