+++ StudentSoup/src/main/frontend/src/components/boards/boardWrite.scss
