+++ StudentSoup/src/main/frontend/src/pages/board/BoardDetail.tsx
@@ -3,8 +3,8 @@ import left from 'assets/images/left.svg';
 import review_white from 'assets/images/review_white.svg';
 import heart_white from 'assets/images/heart_white.svg';
 import heart_border_white from 'assets/images/heart_border_white.svg';
-import BoardBestReview from '../../components/board/BoardBestReview';
-import BoardReview from '../../components/board/BoardReview';
+import BoardBestReview from '../../components/boards/BoardBestReview';
+import BoardReview from '../../components/boards/BoardReview';
 import { Desktop, Mobile } from 'mediaQuery';
 import { useEffect, useState } from 'react';
 import axiosInstance from 'apis/utils/AxiosInterceptor';
